Move gui websocket to subfolder, compile as object file
