+++ gui/src/websocket_server.cpp
@@ -1,0 +1,5 @@
+/**
+ * @brief Make sure to compile an object file.
+ */
+
+#include "gui/websocket_server.hpp"
